improve gc

+++ backr2.py
@@ -229,15 +229,23 @@ def gc(lbh):
         if os.path.exists(lbh + "/store/" + i + "/reference"):
             reference = file_read(lbh + "/store/" + i + "/reference")
             if int(reference) < latest:
-                if verbose:
-                    print("Deleting file " + lbh + "/store/" + i + "/file")
-                os.remove(lbh + "/store/" + i + "/file")
+                if os.path.exists(lbh + "/store/" + i + "/file"):
+                    if verbose:
+                        print("Deleting file " + lbh + "/store/" + i + "/file")
+                    os.remove(lbh + "/store/" + i + "/file")
+                    os.remove(lbh + "/store/" + i + "/reference")
+                if not os.path.exists(lbh + "/store/" + i + "/file.tar.gz"):
+                    shutil.rmtree(lbh + "/store/" + i)
         if os.path.exists(lbh + "/store/" + i + "/reference_tar"):
             reference_tar = file_read(lbh + "/store/" + i + "/reference_tar")
             if int(reference_tar) < latest:
-                if verbose:
-                    print("Deleting file " + lbh + "/store/" + i + "/file.tar.gz")
-                os.remove(lbh + "/store/" + i + "/file.tar.gz")
+                if os.path.exists(lbh + "/store/" + i + "/file.tar.gz"):
+                    if verbose:
+                        print("Deleting file " + lbh + "/store/" + i + "/file.tar.gz")
+                    os.remove(lbh + "/store/" + i + "/file.tar.gz")
+                    os.remove(lbh + "/store/" + i + "/reference_tar")
+                if not os.path.exists(lbh + "/store/" + i + "/file"):
+                    shutil.rmtree(lbh + "/store/" + i)
 
 def compress(filename, outputfile):
     with tarfile.open(outputfile, "w:gz") as tar:
